refactor: move ShiftSchedulePage
